azure_rag_eval/requirements.txt
azure-search-documents==11.4.0
azure-ai-evaluation==1.0.0
azure-identity==1.17.1
openai==1.35.0
python-dotenv==1.0.0
pandas==2.2.2
tqdm==4.66.4
